Analytics Design to Recommend Flavors to Launch
A private label recently launched six flavors--Blueberry, Honey, Peach, Plain, Strawberry, and Vanilla. Our team provided analytics to help determine what the next flavors to launch should be. We had access to the data from a survey that captures information about consumers’ flavor preferences, as well as data pulled from the loyalty card system that the data architect created for the project. 
We used MySQL to describe the percentage of sales of existing flavors in the Greek yogurt category and the rest of the market. Then used R to conduct data cleaning on survey data, also used TURF analysis to calculate the aggregated reach of flavor mix.