Conjoint Analysis - Toy Horse Project
We were in charge of analyzing a conjoint analysis conducted for a relatively small toy company, EarlyRiders, that specializes in the manufacture of quality toy riding horses with a particular focus on natural materials. The dataset available was consumer ratings on different 12 product profiles, with a different mix of 4 attributes: price, height, motion, and style. 
We did data cleaning and dealt with NA firstly. After that, we created a data frame and fill each individual's part-utilities on 4 attributes and predicted missing values (rating). We choose 3 clusters as the best number of customers by elbow rule and visualization techniques also plotted a k-means cluster as three plot reports: pie chart with membership percentages, ellipse plot that indicates cluster definitions against principle components, and barplot of the cluster means.
In section C, we introduced segmentation by demographic variables age and gender. We conducted a significance test by interacting them with product attributes in linear regressions.
In section D, we create 7 product mix scenarios and calculated market share with consideration of competitive impact and response. Lastly, we make a recommendation on which product to launch based on short term and long term profit.